% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiffBetweenEstimates.R
\name{DiffBetweenEstimates}
\alias{DiffBetweenEstimates}
\title{Dfference between estimated parameters.}
\usage{
DiffBetweenEstimates(estimates)
}
\arguments{
\item{estimates}{\code{\link{list}} with two objects of class \code{mcmc.list}. These objects must have the same number of chains and the chains must have the same length.}
}
\value{
a list of class \code{mcmc.list}.
}
\description{
Vectorized difference between chains of two objects of \code{\link{class}} \code{mcmc.list}.
}
\examples{
# Priors.
priors <- c(true.prev.a = 1, true.prev.b = 1,
            se.a = 6.28, se.b = 13.32, sp.a = 212.12, sp.b = 3.13)

# First estimate.
dataset1 <- list(pop.size = 100, positives = 5)
prev.est1 <- OneTestOnePopBM(dataset = as.list(dataset1), n.iter = 3e3,
                                  priors = priors, pars = 'true.prev',
                                  burn.in = 5e2)

# Second estimate.
dataset2 <- list(pop.size = 91, positives = 1)
prev.est2 <- OneTestOnePopBM(dataset = as.list(dataset2), n.iter = 3e3,
                                   priors = priors, pars = 'true.prev',
                                   burn.in = 5e2)

# Estimated difference.
diffs <- DiffBetweenEstimates(list(prev.est1, prev.est2))
summary(diffs)

# Diagnostic plots.
library(coda); library(ggmcmc)
gelman.diag(diffs)
gelman.plot(diffs)
gg.res <- ggs(diffs)
ggs_traceplot(gg.res)
ggs_density(gg.res)
ggs_histogram(gg.res, bins = 100)
ggs_compare_partial(gg.res)
ggs_running(gg.res)
ggs_autocorrelation(gg.res)
}

