% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ElicitBeta.R
\name{ElicitBeta}
\alias{ElicitBeta}
\title{Parameterization of a Beta distribution.}
\usage{
ElicitBeta(mode, confidence, maximum = NULL, minimum = NULL,
  summary = FALSE, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\arguments{
\item{mode}{scalar representing the elicited mode of the variable of interest. Must be a number between 0 and 1.}

\item{confidence}{scalar representing the the level of confidence on the definition of \code{maximum} or \code{minimumt}. Must be a number between 0 and 1.}

\item{maximum}{scalar representing the elicited maximum value of the variable of interest. Must be a number between 0 and 1.}

\item{minimum}{scalar representing the elicited minimum value of the variable of interest. Must be a number between 0 and 1.}

\item{summary}{logical. If FALSE (default), only the parameters a (shape1) and b (shape2) are returned. If TRUE, the mean the variance and the quantiles are also returned.}

\item{quantiles}{\code{\link{numeric}} \code{\link{vector}} with the quantiles to be computed if \code{summary == TRUE}. The quantiles must be numbers between 0 and 1.}
}
\value{
If \code{summary == FALSE}, a \code{\link{numeric}} \code{\link{vector}} with the prameters a and b. If \code{summary == TRUE}, a \code{\link{list}} with the parameters a and b, as well as summary statistics of the distribution.
}
\description{
Parameterization of a Beta distribution from elicited mode and quantiles.
}
\details{
The parameters a and b are calculated using the algorithm implemented in the \code{epi.betabuster} function of the epiR package.
}
\examples{
ElicitBeta(mode = 0.3, maximum = 0.5, confidence = 0.95, summary = FALSE)
ElicitBeta(mode = 0.99, minimum = 0.97, confidence = 0.95, summary = TRUE)
}

