% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneTestMultPopBMM.R
\name{OneTestMultPopBMM}
\alias{OneTestMultPopBMM}
\title{1 test and multiple populations binomial mixture model}
\usage{
OneTestMultPopBMM(dataset, inits, priors, pars, n_iter = 10000,
  n_chains = 3, burn_in = 1000, thin = 1)
}
\arguments{
\item{dataset}{\code{\link{list}} with the population size and the number of positives. The names of these values must be "pop_size" and "positives" respectively.}

\item{inits}{\code{\link{list}} with initial conditions for chains. \code{inits} must define initial values for the true prevalence within infected herds, the sensitivity, the specificity and the herd prevalence. The names of these values must be "true_prev_wph", "se", "sp" and "prev_h" respectively.}

\item{priors}{\code{\link{vector}} with the parameters a and b (Beta distribution) for the true prevalence within infected herds, the sensitivity and the specificity; and with the parameter p (Bernoulli distribution) for the herd prevalence. The names of these values must be: "true_prev_wph_mean_a", "true_prev_wph_mean_b", "se_a", "se_b", "sp_a", "sp_b" "prev_h_a", "prev_h_b", true_prev_wph_var_a", "true_prev_wph_var_b".}

\item{pars}{character vector giving the names of parameters to be monitored. It is passed to the \code{variable.names} argument of the \code{\link{coda.samples}} function. In addition to the parameters specified in \code{inits}, the apparent prevalence, the true prevalence, and the positive and negative predictive values can be monitored if specified in \code{pars} as "app_prev", "true_prev", "ppv" and "npv", respectively.}

\item{thin}{thinning interval for monitors. It is passed to the \code{thin} argument of the \code{\link{coda.samples}} function.}

\item{n.chains}{the number of parallel chains for the model. It is passed to the \code{n.chains} argument of the \code{\link{jags.model}} function.}

\item{burn.in}{the number of iteration to be discarded. It is passed to the \code{n.iter} argument of the \code{\link{update.jags}} function.}

\item{n.iter}{number of iterations to monitor. It is passed to the \code{n.iter} argument of the \code{\link{coda.samples}} function.
#' @details This function creates a text file with the model and it is saved in the working directory.}
}
\value{
A \code{\link{list}} of class \code{mcmc.list}.
}
\description{
1 test and multiple population binomial mixture model to estimate herd prevalence and prevalence within positive herds.
}
\references{
https://dl.dropboxusercontent.com/u/49022/diagnostictests/index.html
}
