% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotElicitedPrior.R
\name{PlotElicitedPrior}
\alias{PlotElicitedPrior}
\title{Plot elicited priors.}
\usage{
PlotElicitedPrior(elicited.prior, ylab = "Density", ...)
}
\arguments{
\item{elicited.prior}{output form the \code{\link{ElicitBeta}} function.}

\item{ylab}{a title for the y axis.}

\item{...}{further arguments passed to the \code{\link{curve}} fuction.}
}
\value{
Density plot.
}
\description{
Plots a Beta distribution parameterized with the \code{\link{ElicitBeta}} function.
}
\examples{
elicited.prior <- ElicitBeta(mode = 0.3, maximum = 0.5, confidence = 0.95)
PlotElicitedPrior(elicited.prior, col = 'red')
}

