% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptb-package.R
\docType{package}
\name{ptb-package}
\alias{ptb-package}
\title{ptb}
\description{
Prevalence estimation and diagnostic test evaluation in a Bayesian framework
}
\details{
\tabular{ll}{
Package: \tab ptb\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
Date: \tab 2018-02-20\cr
Depends: \tab R (>= 3.4)\cr
Imports: \tab rjags\cr
Suggests: \tab coda, ggmcmc\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/leb-fmvz-usp/ptb}\cr
Authors: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{baquero@usp.br}\cr
}
}
\keyword{package}
