% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffBetweenEstimates.R
\name{DiffBetweenEstimates}
\alias{DiffBetweenEstimates}
\title{Difference between estimated parameters}
\usage{
DiffBetweenEstimates(estimates)
}
\arguments{
\item{estimates}{\code{\link{list}} with two objects of class \code{mcmc.list}. These objects must have the same number of chains and the chains must have the same length.}
}
\value{
A list of class \code{mcmc.list}.
}
\description{
Vectorized difference between chains of two objects of \code{\link{class}} \code{mcmc.list}.
}
\examples{

# Priors.
priors <- c(true_prev_a = 1, true_prev_b = 1,
            se_a = 6.28, se_b = 13.32, sp_a = 212.12, sp_b = 3.13)

# First estimate.
dataset1 <- list(pop_size = 100, positives = 5)
prev_est1 <- OneTestOnePopBM(dataset = as.list(dataset1), n_iter = 3e3,
                                  priors = priors, pars = "true_prev",
                                  burn_in = 5e2)

# Second estimate.
dataset2 <- list(pop_size = 91, positives = 1)
prev_est2 <- OneTestOnePopBM(dataset = as.list(dataset2), n_iter = 3e3,
                                   priors = priors, pars = "true_prev",
                                   burn_in = 5e2)

# Estimated difference.
diffs <- DiffBetweenEstimates(list(prev_est1, prev_est2))
summary(diffs)

# Diagnostic plots.
library(coda); library(ggmcmc)
gelman.diag(diffs)
gelman.plot(diffs)
gg_res <- ggs(diffs)
ggs_traceplot(gg_res)
ggs_density(gg_res)
ggs_histogram(gg_res, bins = 100)
ggs_compare_partial(gg_res)
ggs_running(gg_res)
ggs_autocorrelation(gg_res)
}
