% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptb-package.R
\docType{package}
\name{ptb-package}
\alias{ptb-package}
\title{ptb}
\description{
Prevalence estimation and diagnostic test evaluation using Bayesian simulations
}
\details{
\tabular{ll}{
Package: \tab ptb\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-07-12\cr
Depends: \tab R (>= 3.2.1)\cr
Imports: \tab rjags\cr
Suggests: \tab coda, ggmcmc\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/oswaldosantos/ptb}\cr
Authors: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
Maintainer: \tab Oswaldo Santos Baquero \email{oswaldosant@gmail.com}\cr
}
}
\keyword{package}

